Revering headings check to previous code
